commandline ver
commandline ver first
