+++ src/main/java/AppVersion/CommandlineVersion.java
@@ -1,0 +1,5 @@
+package AppVersion;
+
+public class CommandlineVersion {
+
+}
