+++ src/main/java/AppVersion/LibraryAppVersion.java
@@ -1,0 +1,4 @@
+package AppVersion;
+
+public class LibraryAppVersion {
+}
