+++ src/main/java/DocumentManager/Author.java
@@ -1,0 +1,4 @@
+package DocumentManager;
+
+public class Author {
+}
