+++ src/main/java/DocumentManager/Book.java
@@ -1,0 +1,4 @@
+package DocumentManager;
+
+public class Book {
+}
