+++ src/main/java/DocumentManager/DocumentManager.java
@@ -1,0 +1,4 @@
+package DocumentManager;
+
+public class DocumentManager {
+}
