+++ src/main/java/GUI/LibraryDataBase.java
@@ -1,0 +1,5 @@
+package GUI;
+
+public class LibraryDataBase {
+
+}
