+++ src/main/java/GUI/mainStage.java
@@ -1,0 +1,5 @@
+package GUI;
+
+public class mainStage {
+
+}
