+++ src/main/java/User/ReaderAccount.java
@@ -1,0 +1,5 @@
+package User;
+
+public class ReaderAccount extends User {
+
+}
