+++ src/main/java/User/User.java
@@ -1,0 +1,6 @@
+package User;
+
+public class User {
+    private String userId;
+
+}
