--- src/main/java/org/example/library_management/mainStage.java
@@ -1,5 +1,0 @@
-package org.example.library_management;
-
-public class mainStage {
-
-}
